test: :white_check_mark: Added new test and fixed minor bugs related to decoding of binary assignments.
Added end-to-end tests that test the `PathfindingQUBOGenerator` for two specific problem instances on all encodings. These tests have unveiled some bugs in the methods to decode bit arrays into paths, as the ordering of variables was wrong, so this behaviour was updates as well. Also added a feature to allow users to use get_cost on generators that require auxiliary variables even without providing the aux variables themselbes.