:sparkles: Add device-optimized QAOA generation (#158)
## Description

This PR adds new functionalities to the `QuboGenerator` class that allow
for the automated generation of QAOA circuits
- in general, with a new auxiliary variable selection strategy and
without relying on `qiskit_optimization`
- using a hardware-efficient auxiliary variable selection strategy

The second method is explained in detail in the corresponding paper
which is currently under review.

## Checklist:

<!---
This checklist serves as a reminder of a couple of things that ensure
your pull request will be merged swiftly.
-->

- [x] The pull request only contains commits that are related to it.
- [ ] I have added appropriate tests and documentation.
- [ ] I have made sure that all CI jobs on GitHub pass.
- [ ] The pull request introduces no new warnings and follows the
project's style guidelines.

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>